🛠️ Fixed permission error

+++ src/main/java/net/toujoustudios/antithreat/listener/GuildMessageReceivedListener.java
@@ -1,10 +1,15 @@
 package net.toujoustudios.antithreat.listener;
 
+import com.mysql.cj.x.protobuf.Mysqlx;
 import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
+import net.dv8tion.jda.internal.utils.PermissionUtil;
 import net.toujoustudios.antithreat.config.Config;
+import net.toujoustudios.antithreat.error.ErrorEmbed;
+import net.toujoustudios.antithreat.error.ErrorType;
 import net.toujoustudios.antithreat.main.Main;
 import net.toujoustudios.antithreat.util.ColorUtil;
 import org.jetbrains.annotations.NotNull;
@@ -45,12 +50,16 @@ public class GuildMessageReceivedListener extends ListenerAdapter {
                     if (link.contains(site)) {
 
                         EmbedBuilder embedBuilder = new EmbedBuilder();
-                        embedBuilder.setColor(ColorUtil.getFromRGBString(config.getString("format.color.default")));
-                        embedBuilder.setTitle(":warning: **Message Deleted**");
-                        embedBuilder.setDescription("Your message has been deleted for the following reason:\n`Potentially malicious or dangerous links.`\n\n*This message has been automatically reported to administrators.*");
 
-                        event.getMessage().delete().queue();
-                        event.getChannel().sendMessage(embedBuilder.build()).queue();
+                        if(PermissionUtil.checkPermission(event.getChannel(), event.getGuild().getSelfMember(), Permission.MESSAGE_MANAGE)) {
+                            embedBuilder.setColor(ColorUtil.getFromRGBString(config.getString("format.color.default")));
+                            embedBuilder.setTitle(":warning: **Message Deleted**");
+                            embedBuilder.setDescription("Your message has been deleted for the following reason:\n`Potentially malicious or dangerous links.`\n\n*This message has been logged and reported to administrators.*");
+                            event.getMessage().delete().queue();
+                            event.getChannel().sendMessage(embedBuilder.build()).queue();
+                        } else {
+                            ErrorEmbed.sendError(event.getChannel(), ErrorType.PERMISSION_MESSAGE_MANAGE);
+                        }
 
                         embedBuilder.setTitle(":warning: **Scam Link Detection:**");
                         embedBuilder.setDescription("A message was flagged as scam and removed from a server.```" + rawMessage + "```");
